split up main + generate utils
